backend/requirements.txt
fastapi
uvicorn
anthropic
langgraph
langchain-anthropic
langchain-core
python-dotenv
boto3
pydantic